dhcp-client: dhclient to unlink env files
systemctl reload dhclient followed by removal of environment
file assumes that systemctl is synchronous, where dhclient is
invoked asynchronously. Consequently, the environment file may
be removed before dhclient is invoked.

Leave the environment files in place as this forms part of the
dhclient running environment, in a similar fashion to dhcpv6 client.

VRVDR-50952