dhcp-client: start dhcpv6 client on linkup
The netplug notification scripts are processed in the
background (in batches of 3), where run-parts invokes
the scripts in the notificaton directories sequentially.

On interface up, the netplug notifications ordering can
result in newaddr dhcp script overlapping a linkdown dhcp
script notification, where dhcpv6 client fails to start due
to linkdown being actioned last.

Add a dhcpclient-6 script for subsequent linkup notification
to start dhcpv6 client.

The dhcp-client will fail to start on linkup due to check if
the interface exists in /var/lib/dhcp. This file is created by
vyatta-dhcp-client on starting dhcp-client and removed on cleanup.
Replace this check with a systemctl is-active check instead.

VRVDR-51749